feat(app): 'standard' Sponsors added in 'Heroes' section
